Supporting saving state_dict on epoch. Adjust lenghts and saving all data on start.
