+++ bin/lemmatizer_large.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 OUT_DIR=$1
-TRAIN="/home/haukurpj/Resources/Data/MIM-GOLD-SETS.21.05/sets/10TM.tsv /home/haukurpj/Resources/Data/bin_data.tsv"
+TRAIN="/home/haukurpj/Resources/Data/MIM-GOLD-SETS.21.05/sets/10TM.tsv"
 TEST="/home/haukurpj/Resources/Data/MIM-GOLD-SETS.21.05/sets/10PM.tsv"
 
 mkdir -p $OUT_DIR
@@ -22,19 +22,17 @@ train-and-tag \
 $TRAIN \
 "$TEST" \
 "$OUT_DIR" \
---adjust_lengths 1 \
 --lemmatizer \
 --lemmatizer_hidden_dim 512 \
---lemmatizer_state_dict ~/Resources/Models/Lemmatizer/lemmatizer-bin-data-512-trained-1e-4/state_dict_fixed.pt \
 --tag_embedding_dim 128 \
 --char_lstm_layers 1 \
---char_lstm_dim 128 \
---char_emb_dim 64 \
---label_smoothing 0.0 \
+--char_lstm_dim 256 \
+--char_emb_dim 128 \
+--label_smoothing 0.1 \
 --epochs 40 \
---batch_size 512 \
+--batch_size 4 \
 --optimizer adam \
---learning_rate 5e-4 \
+--learning_rate 1e-4 \
 --scheduler none \
 --gpu \
 $*
